+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 This update introduces bug fixes
 
+### Added
+ - Add getter methods for installation lifecycle steps
+
 ### Changed
  - Updated DocStrings in PackageManager